Added support for nPr operator

+++ Datafield/Datafield.cs
@@ -4,6 +4,6 @@ namespace combgen.Datafield;
 
 public abstract class Datafield
 {
-    public abstract DataType Read(int combVal, short? aIndex, short? bIndex);
+    public abstract DataType Read(int combVal, int? aIndex, int? bIndex);
     public abstract int Count();
 }
